Implement the test suit with release apk
The app provides native function jni calling with implementation including warmup, nop and some calculations that can't be optimised out. The app also provides a suit option that executes all the aforementioned tests.
The app is completely user friendly and provides buttons to run a particular test. The output is provided in the text box in the same UI (in nano second).